arm: rename architecture::arm::ap to ap_v1
# Conflicts:
#	probe-rs/src/architecture/arm/communication_interface.rs
#	probe-rs/src/architecture/arm/mod.rs
#	probe-rs/src/vendor/nordicsemi/sequences/nrf53.rs
#	probe-rs/src/vendor/nordicsemi/sequences/nrf91.rs
#	probe-rs/src/vendor/nxp/sequences/nxp_armv8m.rs
