+++ probe-rs-tools/src/bin/probe-rs/cmd/cargo_embed/error.rs
@@ -1,4 +1,4 @@
-use probe_rs::{architecture::arm::ap::AccessPortError, probe::DebugProbeError};
+use probe_rs::{architecture::arm::ap_v1::AccessPortError, probe::DebugProbeError};
 use std::error::Error;
 use std::fmt;
