+++ probe-rs/src/architecture/arm/ap_v1/memory_ap/amba_apb2_apb3.rs
@@ -1,5 +1,5 @@
 use crate::architecture::arm::{
-    ap::{AccessPortType, ApAccess, ApRegAccess},
+    ap_v1::{AccessPortType, ApAccess, ApRegAccess},
     communication_interface::RegisterParseError,
     ArmError, DapAccess, FullyQualifiedApAddress, Register,
 };
@@ -11,13 +11,13 @@ use super::{registers::AddressIncrement, DataSize};
 /// The memory AP can be used to access a memory-mapped
 /// set of debug resources of the attached system.
 #[derive(Debug)]
-pub struct AmbaApb4Apb5 {
+pub struct AmbaApb2Apb3 {
     address: FullyQualifiedApAddress,
     csw: CSW,
     cfg: super::registers::CFG,
 }
 
-impl AmbaApb4Apb5 {
+impl AmbaApb2Apb3 {
     /// Creates a new AmbaAhb3 with `address` as base address.
     pub fn new<P: DapAccess>(
         probe: &mut P,
@@ -40,7 +40,7 @@ impl AmbaApb4Apb5 {
     }
 }
 
-impl super::MemoryApType for AmbaApb4Apb5 {
+impl super::MemoryApType for AmbaApb2Apb3 {
     type CSW = CSW;
 
     fn status<P: ApAccess + ?Sized>(&mut self, probe: &mut P) -> Result<CSW, ArmError> {
@@ -59,7 +59,9 @@ impl super::MemoryApType for AmbaApb4Apb5 {
     ) -> Result<(), ArmError> {
         match data_size {
             DataSize::U32 => Ok(()),
-            _ => Err(ArmError::UnsupportedTransferWidth(data_size as usize * 8)),
+            _ => Err(ArmError::UnsupportedTransferWidth(
+                data_size.to_byte_count() * 8,
+            )),
         }
     }
 
@@ -72,20 +74,20 @@ impl super::MemoryApType for AmbaApb4Apb5 {
     }
 
     fn supports_only_32bit_data_size(&self) -> bool {
-        // APB4 and APB5 AP only support 32bit accesses
+        // APB2 and APB3 AP only support 32bit accesses
         true
     }
 }
 
-impl AccessPortType for AmbaApb4Apb5 {
+impl AccessPortType for AmbaApb2Apb3 {
     fn ap_address(&self) -> &FullyQualifiedApAddress {
         &self.address
     }
 }
 
-impl ApRegAccess<CSW> for AmbaApb4Apb5 {}
+impl ApRegAccess<CSW> for AmbaApb2Apb3 {}
 
-crate::attached_regs_to_mem_ap!(memory_ap_regs => AmbaApb4Apb5);
+crate::attached_regs_to_mem_ap!(memory_ap_regs => AmbaApb2Apb3);
 
 define_ap_register!(
     /// Control and Status Word register
@@ -97,17 +99,6 @@ define_ap_register!(
     fields: [
         /// Is debug software access enabled.
         DbgSwEnable: bool,          // [31]
-        /// Is a Non-secure transfer requested.
-        /// If a secure transfer is requested, the behaviour depends on the value of SPIDEN.
-        /// - If SPIDEN is 1 then a secure transfer is initiated.
-        /// - IF SPIDEN is 0, then no transfer is initiated. An access to DRW or BD0-BD3 is likely
-        ///   to return an error.
-        NonSecure: bool,            // [29]
-        /// Is this transaction privileged
-        Privileged: bool,            // [28]
-        /// May reflect the state of the CoreSight authentication interface.
-        /// If Secure debug is not supported, this field is always 0.
-        SPIDEN: bool,               // [23]
         /// A transfer is in progress.
         /// Can be used to poll whether an aborted transaction has completed.
         /// Read only.
@@ -115,29 +106,24 @@ define_ap_register!(
         /// `1` if transactions can be issued through this access port at the moment.
         /// Read only.
         DeviceEn: bool,             // [6]
-        /// The address increment on DRW access.
+        /// Address Auto Increment.
+        /// This AP does not support the Packed mode of transfer.
         AddrInc: AddressIncrement,  // [5:4]
         /// The access size of this memory AP.
         /// Only supports word accesses.
         Size: DataSize,             // [2:0]
-        /// Reserved
+        /// Reserved bit, kept to preserve IMPLEMENTATION DEFINED statuses.
         _reserved_bits: u32,        // mask
     ],
     from: value => Ok(CSW {
         DbgSwEnable: ((value >> 31) & 0x01) != 0,
-        NonSecure:  ((value >> 29) & 0x01) != 0,
-        Privileged: ((value >> 28) & 0x01) != 0,
-        SPIDEN:     ((value >> 23) & 0x01) != 0,
         TrInProg:   ((value >> 7) & 0x01) != 0,
         DeviceEn:   ((value >> 6) & 0x01) != 0,
         AddrInc: AddressIncrement::from_u8(((value >> 4) & 0x03) as u8).ok_or_else(|| RegisterParseError::new("CSW", value))?,
         Size: DataSize::try_from((value & 0x07) as u8).map_err(|_| RegisterParseError::new("CSW", value))?,
-        _reserved_bits: (value & 0x5F7F_FF08),
+        _reserved_bits: (value & 0x7FFF_FF08),
     }),
     to: value => (u32::from(value.DbgSwEnable) << 31)
-    | (u32::from(value.NonSecure) << 29)
-    | (u32::from(value.Privileged) << 28)
-    | (u32::from(value.SPIDEN) << 23)
     | (u32::from(value.TrInProg) << 7)
     | (u32::from(value.DeviceEn) << 6)
     | ((value.AddrInc as u32) << 4)
