+++ probe-rs/src/architecture/arm/ap_v1/memory_ap/mod.rs
@@ -30,7 +30,7 @@ macro_rules! attached_regs_to_mem_ap {
     ($mod_name:ident => $name:ident) => {
         mod $mod_name {
             use super::$name;
-            use $crate::architecture::arm::ap::{
+            use $crate::architecture::arm::ap_v1::{
                 memory_ap::registers::{
                     BASE, BASE2, BD0, BD1, BD2, BD3, CFG, CSW, DRW, MBT, TAR, TAR2,
                 },
@@ -180,12 +180,12 @@ macro_rules! memory_aps {
                 interface: &mut I,
                 address: &FullyQualifiedApAddress,
             ) -> Result<Self, ArmError> {
-                use crate::architecture::arm::{ap::IDR, Register};
+                use crate::architecture::arm::{ap_v1::IDR, Register};
                 let idr: IDR = interface
                     .read_raw_ap_register(address, IDR::ADDRESS)?
                     .try_into()?;
                 tracing::debug!("reading IDR: {:x?}", idr);
-                use crate::architecture::arm::ap::ApType;
+                use crate::architecture::arm::ap_v1::ApType;
                 Ok(match idr.TYPE {
                     ApType::JtagComAp => return Err(ArmError::WrongApType),
                     $(ApType::$variant => <$type>::new(interface, address.clone())?.into(),)*
