+++ probe-rs/src/architecture/arm/core/armv8a.rs
@@ -1683,7 +1683,7 @@ impl MemoryInterface for Armv8a<'_> {
 mod test {
     use crate::{
         architecture::arm::{
-            ap::memory_ap::MemoryAp, communication_interface::SwdSequence,
+            ap_v1::memory_ap::MemoryAp, communication_interface::SwdSequence,
             sequences::DefaultArmSequence,
         },
         probe::DebugProbeError,
