+++ probe-rs/src/probe/stlink/mod.rs
@@ -6,7 +6,7 @@ mod usb_interface;
 
 use crate::{
     architecture::arm::{
-        ap::{
+        ap_v1::{
             memory_ap::{MemoryAp, MemoryApType},
             valid_access_ports, AccessPortType,
         },
